Encoding twice
I was banging my head against the wall as to why your site was working but not ours. Then I noticed that the first few characters are always the same ('ZG') instead of ('/9'). And turns out the PHP encodes twice base64. Now I don't know if the change above breaks anything else.